third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/cookie-injection-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?alert-cookie=1&q=%3Cmeta%20http-equiv=%22Set-Cookie%22%20content=%22xssAuditorTestCookie=FAIL%22%20/%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
ALERT: PASS
